refactor : 간접참조로 변경
