+++ src/main/java/org/swyp/dessertbee/admin/report/service/MateReportAdminService.java
@@ -1,78 +1,22 @@
 package org.swyp.dessertbee.admin.report.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.swyp.dessertbee.common.exception.BusinessException;
-import org.swyp.dessertbee.common.exception.ErrorCode;
 import org.swyp.dessertbee.community.mate.dto.response.MateReportResponse;
-import org.swyp.dessertbee.community.mate.entity.Mate;
-import org.swyp.dessertbee.community.mate.entity.MateReply;
-import org.swyp.dessertbee.community.mate.entity.MateReport;
-import org.swyp.dessertbee.community.mate.repository.MateReplyRepository;
-import org.swyp.dessertbee.community.mate.repository.MateReportRepository;
-import org.swyp.dessertbee.community.mate.repository.MateRepository;
 
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-public class MateReportAdminService {
-
-    private final MateReportRepository mateReportRepository;
-    private final MateRepository mateRepository;
-    private final MateReplyRepository mateReplyRepository;
+public interface MateReportAdminService {
 
     // 신고된 Mate 게시글 목록 조회
-    public List<MateReportResponse> getReportedMates() {
-        List<MateReport> reports = mateReportRepository.findAllByMateIdIsNotNull();
-        return reports.stream()
-                .map(MateReportResponse::new)
-                .collect(Collectors.toList());
-    }
-
+    List<MateReportResponse> getReportedMates();
 
     //Mate 게시글 삭제
-    @Transactional
-    public void deleteMateByUuid(UUID mateUuid) {
-        Mate mate = mateRepository.findByMateUuidAndDeletedAtIsNull(mateUuid)
-                .orElseThrow(() -> new BusinessException(ErrorCode.MATE_NOT_FOUND));
-
-        // mateUuid가 있는 신고 데이터 확인
-        boolean isReported = mateReportRepository.existsByMateId(mate.getMateId());
-        if (!isReported) {
-            throw new BusinessException(ErrorCode.MATE_NOT_REPORTED);
-        }
-
-        // 게시글 삭제 (soft delete)
-        mate.softDelete();
-        mateRepository.save(mate);
-    }
+    void deleteMateByUuid(UUID mateUuid);
 
     //신고된 Mate 댓글 조회
-    public List<MateReportResponse> getReportedMateReplies() {
-        List<MateReport> reports = mateReportRepository.findAllByMateReplyIdIsNotNull();
-
-        return reports.stream()
-                .map(MateReportResponse::new)
-                .collect(Collectors.toList());
-    }
+    List<MateReportResponse> getReportedMateReplies();
 
     //Mate 댓글 삭제
-    @Transactional
-    public void deleteReportedMateReply(Long mateReplyId) {
-        boolean isReported = mateReportRepository.existsByMateReplyId(mateReplyId);
-        if (!isReported) {
-            throw new BusinessException(ErrorCode.MATE_REPLY_NOT_REPORTED);
-        }
-
-        MateReply mateReply = mateReplyRepository.findByMateReplyIdAndDeletedAtIsNull(mateReplyId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.MATE_REPLY_NOT_FOUND));
-
-        mateReply.softDelete();
-        mateReplyRepository.save(mateReply);
-    }
+    void deleteReportedMateReply(Long mateReplyId);
 }
-
